stream process output as soon as possible
In tasks, it is convenient to get output as soon as possible from
shelled out commands. This also simplifies handling downstream
as only the exit code from commands is relevant.

Adapt the check task to work well in this mode and to show only
the most relevant portions of captured exceptions when errors
are met.